requirements.txt
httpx==0.27.*
beautifulsoup4==4.12.*
python-dotenv==1.0.*
PyYAML==6.*
lxml==5.*
gspread==6.*
google-auth==2.*
pytest==8.*
